Added support for an io.Writer from RingBuffer
Kind of got it working, but I'm not convinced the interface is actually
useful.

Calling the new ring_buffer.writer().write(values) requires []u8 collection
of bytes, so callers need to manually break up larger datatypes into
their constituent bytes as a fixed length array.

Planning to remove this integration after this commit, but wanted to
set it into version control in case we need to bring it back for some
reason.